requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0

# Database
sqlalchemy[asyncio]>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
aiosqlite>=0.19.0
alembic>=1.12.1

# Redis & Cache
redis>=5.0.1
hiredis>=2.2.3

# HTTP Client & External Services
httpx>=0.25.2
aiohttp>=3.9.1

# AWS S3 / Storage
boto3>=1.34.25
botocore>=1.34.25
minio>=7.2.0

# Image Processing
opencv-python>=4.8.1.78
Pillow>=10.1.0
pillow-heif>=0.18.0
python-magic>=0.4.27

# Authentication & Security
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
cryptography>=42.0.0

# Data Validation & Serialization
email-validator>=2.1.0
marshmallow>=3.20.2

# Date/Time
python-dateutil>=2.8.2
pytz>=2023.3

# Configuration & Environment
python-dotenv>=1.0.0

# Logging & Monitoring
structlog>=23.2.0

# ML & AI (Face Recognition & Computer Vision)
torch>=2.0.0
torchvision>=0.15.0
facenet-pytorch>=2.5.3
mtcnn>=0.1.1
numpy>=1.26.0
scipy>=1.11.4
scikit-learn>=1.3.2
scikit-image>=0.22.0

# Utilities
python-multipart>=0.0.6
orjson>=3.9.10
python-json-logger>=2.0.7
tenacity>=8.2.3

# Background Tasks & Scheduling
celery>=5.3.4
dramatiq>=1.15.0
APScheduler>=3.10.4

# Metrics & Monitoring
prometheus-client>=0.19.0
psutil>=5.9.6

# File Processing
chardet>=5.2.0
filetype>=1.2.0


# изменеия
decord>=0.6.0